requirements.txt
aiogram==3.20.0.post0
fastembed==0.6.1
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.19
langchain-text-splitters==0.3.10
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
openai==1.84.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
qdrant-client==1.14.2